Initial support for Flask-Migrate.
DB export and import should _not_ be required. Instead it is necessary to
"db stamp" the production database to make it look like it was created using
migrations.

Running manage.py will nag about running migrations if you haven't done that
already.

set environment variable DB_AUTOMIGRATION to anything truthy to automagically
migrate if you're e.g. inside a container and don't want to manually. Setting
it in the shell script using env is an inelegant solution but seems to work.

fix typo