Update to Ansible 2.2.1 (#558)
Docker-py had to also be updated to 1.7.1. as anything below 1.7 wasn't
supported. ToDo: investigate moving to version 2.x, which is called just
docker in PyPI unless i'm entirely mistaken.

Also took the chance to update sudo->become .

There was something broken in vagrant deployment and the change to the
Vagrantfile **seems** to have fixed it.