support static inventory
Static inventory is useful to cut down deployment dependencies for
Ansible dynamic inventory plugin. It also makes it possible to limit
the hosts returned by the inventory, which is not possible using the
built in OpenStack plugin in Ansible 2.9 [1]

- add static IPs based on network prefix for jump and server hosts
- change main host postfix from 'pebbles' to 'server'
- export public hostname in deployment container environment for
  scripting
- add an option to force Heat stack update
- fix restarting apache 2 in sso container handlers

[1] https://github.com/ansible/ansible/pull/51401

support static inventory

Static inventory is useful to cut down deployment dependencies for
Ansible dynamic inventory plugin. It also makes it possible to limit
the hosts returned by the inventory, which is not possible using the
built in OpenStack plugin in Ansible 2.9 [1]

- add static IPs based on network prefix for jump and server hosts
- change main host postfix from 'pebbles' to 'server'
- export public hostname in deployment container environment for
  scripting
- add an option to force Heat stack update
- fix restarting apache 2 in sso container handlers
- add dcterm.bash to help launching new DCs and shells within the
  existing ones
- remove the option to assign random floating IPs

[1] https://github.com/ansible/ansible/pull/51401