support for k3s deployment
This commit adds support for deploying a K3s based cluster on OpenStack
using Heat. As 'k3s' role is based on Apache licensed code, we have
that separated to a different repo and set the dependency in
requirements.yml. We also bake it in the deployment container so that we
don't depend on the installation at runtime.

The main use case for K3s is to run user workloads on it while the
server application can runs on a separate, shared cluster. Application
Helm charts are modified so that we can also deploy Pebbles on K3s, and
use node selectors to run it only on master node while user workloads
are always run on worker nodes.

Changes:
- add openshift library dependency for ansible k8s module
- Helm charts
  - add node selectors
  - add network policies for access control
  - add kubeconfig secret for remote kubernetes driver support
  - add service account role binding for local admin access for
    namespace creation
- add Heat template for a cluster with a single master and multiple
  nodes
- add custom Traefik config for setting up certificate
- change pebbles_network_prefix to network_prefix, add backwards
  compatibility
- add playbooks for deploying K3s
- unify playbook naming by renaming gitlab_runner.yaml
- add requirements.yml for k3s role, taken from Allas and set up ansible
  galaxy roles path in ansible config
- add a role for label based volume formatting and mounting