k3s: cloud config and server group
Add cloud config/cloud-init configuration option to all VMs in k3s
stack. This is useful for example for configuring 'cloud-user' as the
default user when using an image that has a different convention.

Add anti-affinity server group to compute nodes to minimize the impact
of a single hypervisor failure.

Remove references to permanent_volumes, k3s deployment does not have
other options.