release-5: database backups
Add support for taking backups from database and uploading them to
object storage. This is done by setting a K8s CronJob to `pg_dump` the
database contents through `gpg` to a backup volume and using `s3cmd` to
synchronize the backups to object storage. Pebbles-deployer image is
used for the job.

Changes:
- Helm
  - add backup cronjob
  - add backup volume
  - add backup secret
- pebbles-deployer
  - add postgresql client, epel and s3cmd
  - change permissions so that it runs on OpenShift
  - define HOME so that `gpg` is happy out of the box
- add build functions and OpenShift image objects