helm: define starting deadline for backup job
Define startingDeadlineSeconds for backup cronjob to mark the backup job
as failed and start again in case the job does not start timely. This
could happen during maintenance, for example. Without this, the job
can get stuck with error 'Cannot determine if job needs to be started:
Too many missed start time'