add k3s-autoscaler
Add k3s-autoscaler container, a pipeline task to build the container
and a playbook to deploy it on a cluster jump host.

The Autoscaler works for K3s on OpenStack. It watches the workloads on
a K3s cluster and scales the cluster up and down by provisioning and
deleting node VMs in OpenStack. Nodes are supposed to be configured
with Ignition/Butane - in practice Fedora CoreOS fits the bill.

The playbook installs the autoscaler as a systemd unit that runs
the container with podman.