Make default network policy configurable
The allowed namespaces for ingress traffic for the default
network policy can now be configured in values.yaml.
The default network policy was modified to use the universal
label "kubernetes.io/metadata.name" instead of "name" to configure
allowed ingress traffic. These changes are needed for networking
to work with a microk8s cluster.